qseecom: Increase max number of sg entries to 512
The number of scattered entries returned by ion_sg_table()
for a non-contiguous memory is larger than 10 in cases where the
memory is larger.

This patch fixes issues seen in some clients that issue a large
buffer and the number of scattered segments is large.

The max of 512 is calculated based on the max packet size to not
exceed 2MB (with each segment being the minumum of 4KB).

Change-Id: Id45e0212466eb9cbc260d52054670981d4fc1062
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>