cpufreq: Preserve sysfs files across suspend/resume
The file permissions of cpufreq per-cpu sysfs files are not preserved
across suspend/resume because we internally go through the CPU
Hotplug path which reinitializes the file permissions on CPU online.

But the user is not supposed to know that we are using CPU hotplug
internally within suspend/resume (IOW, the kernel should not silently
wreck the user-set file permissions across a suspend cycle).
Therefore, we need to preserve the file permissions as they are
across suspend/resume.

The simplest way to achieve that is to just not touch the sysfs files
at all - ie., just ignore the CPU hotplug notifications in the
suspend/resume path (_FROZEN) in the cpufreq hotplug callback.

Reported-by: Robert Jarzmik <robert.jarzmik@intel.com>
Reported-by: Durgadoss R <durgadoss.r@intel.com>
Signed-off-by: Srivatsa S. Bhat <srivatsa.bhat@linux.vnet.ibm.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

This fixes the problem on d2-aokpx with hot reboots due to Emergency
R/O Remount.
Since we set universal R/W permissions for cpufreq-related sysfs,
on cpu hotplug suspends/resume the permissions were readjusted. Causing
the remount (on ocassion using sysrq to force it, I believe), which
made the device reboot as defined in fs/namespace for remounts.

Conflicts:

	drivers/cpufreq/cpufreq.c

Signed-off-by: Luis Cruz <ljc2491@gmail.com>