sched-dl: Revert to some older defines
The new ABI for sched-dl is based on the latest Linux mainline kernel.
Seems like most of my building problems are related to a couple of defines
that were simply changed. Solution is simply to revert to something more
compatible with our kernel. This should cause no change in the way sched
deadline works altogether.

Signed-off-by: Luis Cruz <ljc2491@gmail.com>